+++ src/main/java/personality/experience/Experience.java
@@ -1,4 +1,4 @@
-package personality;
+package personality.experience;
 
 public interface Experience {
 }
